Refactor into classes and add scrolling level
